Separate middlewares from active learning core
